[feat][caleb] adds routing for wiki page
