+++ src/app/(account)/layout.tsx
@@ -8,9 +8,11 @@ import { toast } from "react-toastify";
 export default function AccountLayout({
   lawyer,
   layman,
+  children,
 }: {
   lawyer: React.ReactNode;
   layman: React.ReactNode;
+  children: React.ReactNode;
 }) {
   const handleLogout = async () => {
     try {
@@ -52,7 +54,7 @@ export default function AccountLayout({
             name: "Submitted Cases",
             path: (id: string) => `/${id}`,
           },
-          { key: "wiki", name: "Wiki", path: (id: string) => `/${id}` },
+          { key: "wiki", name: "Wiki", path: () => `/wiki` },
           { key: "forum", name: "Forum", path: (id: string) => `/${id}` },
           {
             key: "settings",
@@ -68,7 +70,7 @@ export default function AccountLayout({
             name: "Active Cases",
             path: (id: string) => `/${id}/active-cases`,
           },
-          { key: "wiki", name: "Wiki", path: (id: string) => `/wiki` },
+          { key: "wiki", name: "Wiki", path: () => `/wiki` },
           { key: "forum", name: "Forum", path: (id: string) => `/forum` },
           {
             key: "settings",
@@ -79,6 +81,15 @@ export default function AccountLayout({
 
   const activeTab =
     tabs.find((tab) => pathname === tab.path(userId))?.name || "";
+  
+  const isWikiPage = pathname === '/wiki';
+
+  let contentToShow;
+  if (isWikiPage ) {
+    contentToShow = children;
+  } else {
+    contentToShow = userType === 'layman' ? layman : lawyer;
+  }
 
   const sidebarBg = userType === "lawyer" ? "bg-blue/60" : "bg-red/60";
 
@@ -115,7 +126,7 @@ export default function AccountLayout({
         </button>
       </div>
       <div className="h-[95vh] w-[90vw] bg-white p-10 pt-5 rounded-xl z-10 ml-[23vw] mr-[20px] border-none">
-        {userType === "layman" ? layman : lawyer}
+        {contentToShow}
       </div>
     </div>
   );
